+++ m1well.Expenses/src/expenses.js
@@ -1,24 +1,32 @@
 // @flow
 
-import { getYearMonthDate } from '../../helpers/dateTime'
+import { getMonth, getYear } from 'date-fns'
 import { getOrMakeConfigurationSection } from '../../nmn.Templates/src/configuration'
-import { leftPadWithZeros } from '../../helpers/general'
 import { getInputTrimmed, inputNumber } from '../../helpers/userInput'
 import {
-  aggregateCategoriesPerMonth,
+  aggregateByCategoriesAndMonth,
+  castFixedExpensesArrayFromMixed,
+  castShortcutExpensesArrayFromMixed,
+  castStringArrayFromMixed,
+  castStringFromMixed,
+  createAggregationExpenseRowWithDelimiter,
+  createTrackingExpenseRowWithConfig,
   extractExpenseRowFromCsvRow,
-  extractFixedExpensesArrayFromMixed,
-  extractStringArrayFromMixed,
-  extractStringFromMixed,
-  logError,
-  logMessage
+  stringifyShortcutList
 } from './expensesHelper'
-import type { Config, ExpenseRow } from './expensesModels'
+import type { Config, ExpenseTrackingRow } from './expensesModels'
+import { amountOk, categoryOk, logError, logMessage, validateConfig } from './expensesChecks'
+
+const DEFAULT_DELIMITER = ';'
+const ALLOWED_DELIMTER = [ ';', '%', 'TAB' ]
 
 const CONFIG_KEYS = {
   folderPath: 'folderPath',
+  delimiter: 'delimiter',
+  dateFormat: 'dateFormat',
+  columnOrder: 'columnOrder',
   categories: 'categories',
-  shortcuts: 'shortcuts',
+  shortcutExpenses: 'shortcutExpenses',
   fixedExpenses: 'fixedExpenses',
 }
 
@@ -34,7 +42,21 @@ const EXAMPLE_CONFIG = `
   expenses: {
     // just an example folderPath - please adapt to your needs
     ${CONFIG_KEYS.folderPath}: 'finances',
+    // just an example delimiter - you can use ';', '%' or 'TAB'
+    ${CONFIG_KEYS.delimiter}: '${DEFAULT_DELIMITER}',
+    // custom date format - e.g. one date '2021-12-08', or only year and month as columns '2021;12'
+    // so the format should be like 'yyyy-MM-dd' or 'yyyy-MM' - ATTENTION: don't use your chosen delimiter here
+    ${CONFIG_KEYS.dateFormat}: 'yyyy-MM-dd',
+    // please choose your column order before first tracking
+    // there is no eventlistener in the config - to change existing data after changing the order here
+    ${CONFIG_KEYS.columnOrder}: [
+      'date',
+      'category',
+      'text',
+      'amount',
+    ],
     // just some example categories - please adapt to your needs
+    // and set order to your best experience
     ${CONFIG_KEYS.categories}: [
       'Living',
       'Groceries',
@@ -43,10 +65,24 @@ const EXAMPLE_CONFIG = `
       'Media',
       'Fun',
     ],
-    // just some example shortcuts - please adapt to your needs
-    ${CONFIG_KEYS.shortcuts}: [
-      'Mobility;Refuel',
-      'Groceries;XYZ Market',
+    // just some example shortcut expenses - please adapt to your needs
+    // you can also add an amount, then you can insert the shortcut without any question
+    ${CONFIG_KEYS.shortcutExpenses}: [
+      {
+        category: 'Mobility',
+        text: 'Refuel',
+        amount: null,
+      },
+      {
+        category: 'Groceries',
+        text: 'XYZ Market',
+        amount: null,
+      },
+      {
+        category: 'Fun',
+        text: 'Cofe at Starbucks',
+        amount: 8,
+      },
     ],
     // just some example fixed expenses - please adapt to your needs
     ${CONFIG_KEYS.fixedExpenses}: [
@@ -67,7 +103,7 @@ const EXAMPLE_CONFIG = `
       {
         category: 'Media',
         text: 'Spotify',
-        amount: 9.99,
+        amount: 10,
         month: 0,
         active: false,
       },
@@ -77,38 +113,20 @@ const EXAMPLE_CONFIG = `
 `
 
 /**
- * @description expenses tracking with three posibilities (individual, shortcuts, fixed)
+ * @description expenses tracking with three possibilities (individual, shortcuts, fixed)
+ *
  * @returns {Promise<boolean>}
  */
 const expensesTracking = async (): Promise<boolean> => {
-  const config = await provideConfig()
-  if (!config.folderPath) {
-    return false
-  }
-
-  const { year, month } = getYearMonthDate(new Date())
-
-  if (config.categories.length < 1) {
-    // if there are no categories configured, then stop
-    logError('no categories configured')
-    return false
-  }
-
   const mode = await CommandBar.showOptions(TRACKING_MODE, 'Please choose tracking mode')
 
-  const noteTitle = `${year} Expenses Tracking`
-
-  if (!await provideAndCheckNote(noteTitle, config.folderPath, true)) {
-    return false
-  }
-
   switch (mode.value) {
-    case TRACKING_MODE[0]:
-      return individualTracking(config, year, month, noteTitle)
-    case TRACKING_MODE[1]:
-      return shortcutsTracking(config, year, month, noteTitle)
-    case TRACKING_MODE[2]:
-      return fixedTracking(config, year, month, noteTitle)
+    case 'Individual':
+      return await individualTracking()
+    case 'Shortcuts':
+      return await shortcutsTracking()
+    case 'Fixed':
+      return await fixedTracking()
     default:
       return false
   }
@@ -116,10 +134,13 @@ const expensesTracking = async (): Promise<boolean> => {
 
 /**
  * @description aggregates expenses of given year to a new note
+ *
  * @returns {Promise<boolean>}
  */
 const expensesAggregate = async (): Promise<boolean> => {
-  const config = await provideConfig()
+  let config = await provideConfig()
+  config = validateConfig(config, new Date(), DEFAULT_DELIMITER, ALLOWED_DELIMTER)
+
   if (!config.folderPath) {
     return false
   }
@@ -132,29 +153,29 @@ const expensesAggregate = async (): Promise<boolean> => {
   }
 
   await Editor.openNoteByTitle(noteTitleTracking)
-  let trackingData = Editor.paragraphs
+  const trackedData = Editor.paragraphs
     .filter(para => !para.rawContent.startsWith('#'))
-    .map(para => extractExpenseRowFromCsvRow(para.rawContent))
+    .map(para => extractExpenseRowFromCsvRow(para.rawContent, config))
 
-  if (!checkDataQualityBeforeAggregate(trackingData, year, config)) {
+  if (!checkDataQualityBeforeAggregate(trackedData, year, config)) {
     return false
   }
 
-  trackingData = aggregateCategoriesPerMonth(trackingData)
+  const aggregatedData = aggregateByCategoriesAndMonth(trackedData, config.delimiter)
 
   const noteTitleAggregate = `${year} Expenses Aggregate`
   if (!await provideAndCheckNote(noteTitleAggregate, config.folderPath, true)) {
     return false
   }
 
-  if (trackingData.length > 0) {
+  if (aggregatedData.length > 0) {
     await Editor.openNoteByTitle(noteTitleAggregate)
     Editor.removeParagraphs(Editor.paragraphs
       .filter(para => !para.rawContent.startsWith('#')))
     // add results
-    trackingData.forEach(res => {
-      if (res.year) {
-        const line = `${year};${leftPadWithZeros(res.month, 2)};${res.category};${Math.round(res.amount)}`
+    aggregatedData.forEach(aggregated => {
+      if (aggregated.year) {
+        const line = createAggregationExpenseRowWithDelimiter(aggregated, config.delimiter)
         Editor.appendParagraph(line, 'text')
       }
     })
@@ -166,57 +187,151 @@ const expensesAggregate = async (): Promise<boolean> => {
 }
 
 /**
- * @private
+ * @description tracking of individual expenses
+ *
+ * @returns {Promise<boolean>}
  */
-const individualTracking = async (config: Config, year: number, month: number, noteTitle: string): Promise<boolean> => {
+const individualTracking = async (): Promise<boolean> => {
+  const currentDate = new Date()
+  let config = await provideConfig()
+  config = validateConfig(config, currentDate, DEFAULT_DELIMITER, ALLOWED_DELIMTER)
+
+  if (!config.folderPath) {
+    return false
+  }
+
+  const title = `${getYear(currentDate)} Expenses Tracking`
+
+  if (!await provideAndCheckNote(title, config.folderPath, true)) {
+    return false
+  }
+
   const category = await CommandBar.showOptions(config.categories, 'Please choose category')
   const text = await getInputTrimmed('Please type in some text (no semicolon)', 'Add text to expenses line')
   const amount = await inputNumber('Please type in amount (only integer numbers)')
 
-  if (!category || !text || isNaN(amount)) {
+  if (!amountOk(amount)) {
+    logError('amount too big or not a number')
+    return false
+  }
+
+  if (!category || !text) {
     // if user missed some input, then stop
     logError('an input was missing')
     return false
   }
 
-  await Editor.openNoteByTitle(noteTitle)
-  const line = `${year};${leftPadWithZeros(month, 2)};${category.value};${text};${Math.round(amount)}`
+  await Editor.openNoteByTitle(title)
+  const expenseRow = {
+    date: currentDate,
+    category: category.value,
+    text: text,
+    amount: Math.round(amount),
+  }
+  const line = createTrackingExpenseRowWithConfig(expenseRow, config)
   Editor.appendParagraph(line, 'text')
 
   return true
 }
 
 /**
- * @private
+ * @description tracking of shortcut expenses
+ *
+ * @returns {Promise<boolean>}
  */
-const shortcutsTracking = async (config: Config, year: number, month: number, noteTitle: string): Promise<boolean> => {
-  const shortcut = await CommandBar.showOptions(config.shortcuts, 'Please choose shortcut')
-  const amount = await inputNumber('Please type in amount (only integer numbers)')
+const shortcutsTracking = async (): Promise<boolean> => {
+  const currentDate = new Date()
+  let config = await provideConfig()
+  config = validateConfig(config, currentDate, DEFAULT_DELIMITER, ALLOWED_DELIMTER)
 
-  const splittedShortcut = shortcut.value.split(';')
-
-  if (!splittedShortcut[0] || !splittedShortcut[1] || isNaN(amount)) {
-    // if user missed some input, then stop
-    logError('an input was missing')
+  if (!config.folderPath) {
     return false
   }
 
-  await Editor.openNoteByTitle(noteTitle)
-  const line = `${year};${leftPadWithZeros(month, 2)};${splittedShortcut[0]};${splittedShortcut[1]};${Math.round(amount)}`
+  const title = `${getYear(currentDate)} Expenses Tracking`
+
+  if (!await provideAndCheckNote(title, config.folderPath, true)) {
+    return false
+  }
+
+  const shortcut = await CommandBar.showOptions(stringifyShortcutList(config.shortcutExpenses, config.delimiter), 'Please choose shortcut')
+  const selected = config.shortcutExpenses[shortcut.index]
+
+  let amount = 0
+  if (!selected.amount) {
+    amount = await inputNumber('Please type in amount (only integer numbers)')
+  } else {
+    amount = selected.amount
+  }
+
+  if (!amountOk(amount)) {
+    logError('amount not in range or not a number')
+    return false
+  }
+
+  if (!categoryOk(selected.category, config.categories)) {
+    logError('category not configured')
+    return false
+  }
+
+  if (!selected.text) {
+    // if there was no text in the shortcut, then stop
+    logError('text was missing')
+    return false
+  }
+
+  await Editor.openNoteByTitle(title)
+  const expenseRow = {
+    date: currentDate,
+    category: selected.category,
+    text: selected.text,
+    amount: Math.round(amount),
+  }
+  const line = createTrackingExpenseRowWithConfig(expenseRow, config)
   Editor.appendParagraph(line, 'text')
 
   return true
 }
 
 /**
- * @private
+ * @description tracking of fixed expenses
+ *
+ * @returns {Promise<boolean>}
  */
-const fixedTracking = async (config: Config, year: number, month: number, noteTitle: string): Promise<boolean> => {
-  await Editor.openNoteByTitle(noteTitle)
+const fixedTracking = async (): Promise<boolean> => {
+  const currentDate = new Date()
+  let config = await provideConfig()
+  config = validateConfig(config, currentDate, DEFAULT_DELIMITER, ALLOWED_DELIMTER)
+
+  if (!config.folderPath) {
+    return false
+  }
+
+  const title = `${getYear(currentDate)} Expenses Tracking`
+
+  if (!await provideAndCheckNote(title, config.folderPath, true)) {
+    return false
+  }
+
+  const month = getMonth(currentDate) - 1
+
+  await Editor.openNoteByTitle(title)
   config.fixedExpenses
-    .filter(exp => exp.active && (exp.month === 0 || exp.month === Number(month)))
+    .filter(exp => exp.active && (exp.month === 0 || exp.month === month))
+    .map(exp => {
+      if (!categoryOk(exp.category, config.categories)) {
+        exp.category = `>>WRONG CATEGORY (${exp.category})<<`
+      }
+      return exp
+    })
     .forEach(exp => {
-      const line = `${year};${leftPadWithZeros(month, 2)};${exp.category};${exp.text};${Math.round(exp.amount)}`
+      const expenseRow = {
+        date: currentDate,
+        category: exp.category,
+        text: exp.text,
+        amount: Math.round(exp.amount),
+      }
+      const line = createTrackingExpenseRowWithConfig(expenseRow, config)
       Editor.appendParagraph(line, 'text')
     })
 
@@ -224,6 +339,8 @@ const fixedTracking = async (config: Config, year: number, month: number, noteTi
 }
 
 /**
+ * @description provide config from _configuration and cast content to real objects
+ *
  * @private
  */
 const provideConfig = (): Promise<Config> => {
@@ -233,20 +350,26 @@ const provideConfig = (): Promise<Config> => {
   )
     .then(result => {
       if (result == null || Object.keys(result).length === 0) {
-        logError('exptected config could not be found in the _configuration file')
+        logError('expected config could not be found in the _configuration file')
         return {
           folderPath: '',
+          delimiter: '',
+          dateFormat: '',
+          columnOrder: [],
           categories: [],
-          shortcuts: [],
+          shortcutExpenses: [],
           fixedExpenses: []
         }
       } else {
         logMessage(`loaded config\n${JSON.stringify(result)}\n`)
         const config: Config = {
-          folderPath: extractStringFromMixed(result, CONFIG_KEYS.folderPath),
-          categories: extractStringArrayFromMixed(result, CONFIG_KEYS.categories),
-          shortcuts: extractStringArrayFromMixed(result, CONFIG_KEYS.shortcuts),
-          fixedExpenses: extractFixedExpensesArrayFromMixed(result, CONFIG_KEYS.fixedExpenses),
+          folderPath: castStringFromMixed(result, CONFIG_KEYS.folderPath),
+          delimiter: castStringFromMixed(result, CONFIG_KEYS.delimiter),
+          dateFormat: castStringFromMixed(result, CONFIG_KEYS.dateFormat),
+          columnOrder: castStringArrayFromMixed(result, CONFIG_KEYS.columnOrder),
+          categories: castStringArrayFromMixed(result, CONFIG_KEYS.categories),
+          shortcutExpenses: castShortcutExpensesArrayFromMixed(result, CONFIG_KEYS.shortcutExpenses),
+          fixedExpenses: castFixedExpensesArrayFromMixed(result, CONFIG_KEYS.fixedExpenses),
         }
         return config
       }
@@ -254,13 +377,15 @@ const provideConfig = (): Promise<Config> => {
 }
 
 /**
+ * @description check if one note exists by name, if mulitple exists - throw error, of none extist -> create it
+ *
  * @private
  */
-const provideAndCheckNote = async (noteTitle: string,
+const provideAndCheckNote = async (title: string,
                                    folderPath: string,
                                    createNote: boolean,
                                    year?: number): Promise<boolean> => {
-  const notes = DataStore.projectNoteByTitle(noteTitle)
+  const notes = DataStore.projectNoteByTitle(title)
 
   // create note if it de
   if (notes) {
@@ -271,7 +396,7 @@ const provideAndCheckNote = async (noteTitle: string,
     }
     if (notes.length < 1) {
       if (createNote) {
-        await DataStore.newNote(noteTitle, folderPath)
+        await DataStore.newNote(title, folderPath)
         return true
       } else {
         // if there is no note to aggregate
@@ -289,29 +414,30 @@ const provideAndCheckNote = async (noteTitle: string,
 }
 
 /**
+ * @description check data quality of tracked data before we aggregate it
+ *
  * @private
  */
-const checkDataQualityBeforeAggregate = (rows: ExpenseRow[], year: number, config: Config): boolean => {
+const checkDataQualityBeforeAggregate = (rows: ExpenseTrackingRow[], year: number, config: Config): boolean => {
   for (const row of rows) {
-    if (row.year !== year) {
-      logError(`year at: ${leftPadWithZeros(row.month, 2)};${row.category};${row.text ?? ''};${row.amount}`)
+    const rowYear = getYear(row.date)
+
+    if (rowYear !== year) {
+      logError(`year at: ${createTrackingExpenseRowWithConfig(row, config)}`)
       return false
     }
-    if (Number(row.month) < 1 || Number(row.month) > 12) {
-      logError(`month at: ${leftPadWithZeros(row.month, 2)};${row.category};${row.text ?? ''};${row.amount}`)
+    if (!categoryOk(row.category, config.categories)) {
+      logError(`category not found at: ${createTrackingExpenseRowWithConfig(row, config)}`)
       return false
     }
-    if (!row.category || config.categories.findIndex(cl => cl === row.category) === -1) {
-      logError(`category not found at: ${leftPadWithZeros(row.month, 2)};${row.category};${row.text ?? ''};${row.amount}`)
+    if (!amountOk(row.amount)) {
+      logError(`amount at: ${createTrackingExpenseRowWithConfig(row, config)}`)
       return false
     }
-    if (isNaN(row.amount) || row.amount < 0) {
-      logError(`amount at: ${leftPadWithZeros(row.month, 2)};${row.category};${row.text ?? ''};${row.amount}`)
-      return false
-    }
+
   }
 
   return true
 }
 
-export { expensesTracking, expensesAggregate }
+export { expensesTracking, expensesAggregate, individualTracking, shortcutsTracking, fixedTracking }